Put app contents inside grid
