+++ src/App.vue
@@ -1,12 +1,17 @@
 <template>
   <div id="app">
-    <Header />
-    <AttributeSelection />
-    <TimeSpanSelection />
-    <SiteSelection />
-    <DataDownload />
-    <DataPreview />
-    <Footer />
+    <Header id="site-header" />
+    <SiteImage id="site-image" />
+    <SiteTitle id="site-title" />
+    <div id="content">
+      <ErrorMessages />
+      <AttributeSelection />
+      <TimeSpanSelection />
+      <SiteSelection />
+      <DataDownload />
+      <DataPreview />
+    </div>
+    <Footer id="footer" />
     <div id="busy-indicator" v-if="loading">
       <div class="spinner">
         <div class="double-bounce1"></div>
@@ -14,9 +19,7 @@
       </div>
       {{ $t("$busy") }}
     </div>
-    <div v-if="hasError" class="error-notification">
-      {{ $t("$errorNotification") }}
-    </div>
+    <div v-if="hasError" class="error-notification">{{ $t("$errorNotification") }}</div>
   </div>
 </template>
 
@@ -32,11 +35,35 @@
 }
 
 #app {
+  display: grid;
   font-family: "TitilliumWeb", sans-serif;
   text-align: center;
-  max-width: 65rem;
-  margin-left: auto;
-  margin-right: auto;
+  grid-template:
+    "head head head"
+    "pict pict pict"
+    ". cont ."
+    "foot foot foot"
+    / auto 65rem auto;
+}
+
+#site-header {
+  grid-area: head;
+}
+
+#site-image {
+  grid-area: pict;
+}
+
+#site-title {
+  grid-area: pict;
+}
+
+#content {
+  grid-area: cont;
+}
+
+#footer {
+  grid-area: foot;
 }
 
 .hidden {
