+++ src/assets/styles/variables.scss
@@ -25,8 +25,8 @@ $border-red: #C81740;
 $border-divide: #C9CDD1;
 
 //font settings
-$font-size-xxl: 2.7rem;
+$font-size-xxl: 2.625rem;
 $font-size-xl: 1.9rem;
-$font-size-l: 1.2rem;
+$font-size-l: 1.25rem;
 $font-size-m: 1rem;
 $font-size-s: 0.9rem;
