+++ src/components/SiteTitle.vue
@@ -1,0 +1,33 @@
+<template>
+  <header>
+    <h1 id="site-title">{{ $t('$siteTitle')}}</h1>
+    <h2 id="title-details">{{ $t('$siteDetails')}}</h2>
+  </header>
+</template>
+
+<script lang="ts">
+import { Component, Vue } from 'vue-property-decorator';
+import i18n from '@/locale/i18n';
+
+@Component({ i18n })
+export default class SiteTitle extends Vue { }
+</script>
+
+<style lang="scss" scoped>
+@import "@/assets/styles/variables.scss";
+
+#site-title {
+  font-family: "TitilliumWeb";
+  font-size: $font-size-xxl;
+  font-weight: bold;
+  color: $text-white;
+  padding-top: 5%;
+}
+
+#title-details {
+  font-family: "TitilliumWeb";
+  font-size: $font-size-l;
+  font-weight: normal;
+  color: $text-white;
+}
+</style>
